+++ hooks/api/product/index.ts
@@ -1,10 +1,6 @@
 import { queryKey } from '@constants/react-query';
 import { isAxiosError } from 'api/core';
-import {
-  deleteProductDetail,
-  getProductDetail,
-  updateProductDetail,
-} from 'api/product';
+import { deleteProductDetail, getProductDetail } from 'api/product';
 import { queryClient } from 'pages/_app';
 import { toastError, toastSuccess } from 'utils/toaster';
 
@@ -23,23 +19,6 @@ export function useDeleteProduct(id: string) {
       queryClient.removeQueries(queryKey.productDetail(id));
       queryClient.invalidateQueries(['productItemList']);
       toastSuccess({ message: '상품을 삭제했습니다.' });
-    },
-    onError: (err) => {
-      if (isAxiosError<res.ProductDeleteError>(err) && !!err.response) {
-        const { message } = err.response.data;
-        toastError({ message });
-      }
-    },
-  });
-  return response;
-}
-
-export function useUpdateProduct(id: string) {
-  const response = useCoreMutation(updateProductDetail, {
-    onSuccess: () => {
-      queryClient.invalidateQueries(queryKey.productDetail(id));
-      queryClient.invalidateQueries(['productItemList']);
-      toastSuccess({ message: '상품을 수정했습니다.' });
     },
     onError: (err) => {
       if (isAxiosError<res.ProductDeleteError>(err) && !!err.response) {
