+++ hooks/api/upload/index.ts
@@ -4,8 +4,13 @@ import { useCallback } from 'react';
 
 import { queryKey } from '@constants/react-query';
 import { isAxiosError } from 'api/core';
-import { getUploadedProduct } from 'api/product';
-import { postImgs, postProduct } from 'api/upload';
+import {
+  postImgs,
+  postProduct,
+  updateProductDetail,
+  getUploadedProduct,
+} from 'api/upload';
+import { queryClient } from 'pages/_app';
 import { useUploadStore } from 'store/upload/useUploadStore';
 import { toastError, toastSuccess } from 'utils/toaster';
 
@@ -16,13 +21,13 @@ export const useImgUpload = () => {
 };
 
 export const useProductUpload = () => {
-  const clearMeasure = useUploadStore(
-    useCallback((state) => state.clearMeasure, []),
+  const clearUpload = useUploadStore(
+    useCallback((state) => state.clearUpload, []),
   );
   return useCoreMutation(postProduct, {
     onSuccess: ({ data }) => {
       router.push(`/shop/${data}`);
-      clearMeasure();
+      clearUpload();
       toastSuccess({ message: '상품 등록에 성공했습니다.' });
     },
     onError: (err) => {
@@ -33,6 +38,24 @@ export const useProductUpload = () => {
     },
   });
 };
+
+export function useUpdateProduct(id: string) {
+  const response = useCoreMutation(updateProductDetail, {
+    onSuccess: ({ data }) => {
+      router.push(`/shop/${data}`);
+      queryClient.invalidateQueries(queryKey.productDetail(id));
+      queryClient.invalidateQueries(['productItemList']);
+      toastSuccess({ message: '상품을 수정했습니다.' });
+    },
+    onError: (err) => {
+      if (isAxiosError<res.ProductDeleteError>(err) && !!err.response) {
+        const { message } = err.response.data;
+        toastError({ message });
+      }
+    },
+  });
+  return response;
+}
 
 export const useUploadedProduct = (id: string) => {
   return useCoreQuery(queryKey.uploadedProduct(id), () =>
